TECH-3769 metrics for latest and s3: Extract a module; add a comment
